+++ docs/changelog.md
@@ -2,6 +2,10 @@
 
 Here are the most notable changes in each release. For a more detailed list of changes, see the [Github Releases page](https://github.com/heyman/heynote/releases).
 
+## main (next release)
+
+- Prevent blocks from unfolding when changing the block language
+
 ## 2.3.2
 
 - Improve folding to be less aggressive with the unfolding (e.g. when creating a new block after a folded block, and then immediately pressing backspace to remove the new block)